Improve ayo quest 1.1.8

+++ src/main/java/br/com/ayo_quest/ayo_quest/dto/ModuloCadastroDTO.java
@@ -13,7 +13,7 @@ public class ModuloCadastroDTO {
 
     private Long cargaHoraria;
 
-    private Integer xpAoConcluir;
+    private Long xpAoConcluir;
 
     private TrilhaDTO trilha;
